requirements.txt
mediapipe==0.10.21
numpy==2.3.0
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
Pillow==11.2.1
streamlit==1.31.1
streamlit_drawable_canvas==0.9.3
